Move model card getting into a service
