+++ backend/src/routes/v2/model/modelcard/getModelCard.ts
@@ -3,13 +3,8 @@ import { Request, Response } from 'express'
 import { z } from 'zod'
 
 import { ModelCardInterface } from '../../../../models/v2/Model.js'
-import { getModelById, getModelCardRevision } from '../../../../services/v2/model.js'
-import { NotFound } from '../../../../utils/v2/error.js'
+import { getModelCard as getModelCardService, GetModelCardVersionOptions } from '../../../../services/v2/model.js'
 import { parse } from '../../../../utils/validate.js'
-
-export const GetModelCardVersionOptions = {
-  Latest: 'latest',
-} as const
 
 export const getModelCardSchema = z.object({
   params: z.object({
@@ -31,18 +26,7 @@ export const getModelCard = [
       params: { modelId, version },
     } = parse(req, getModelCardSchema)
 
-    let modelCard: ModelCardInterface
-    if (version === GetModelCardVersionOptions.Latest) {
-      const card = (await getModelById(req.user, modelId)).card
-
-      if (!card) {
-        throw NotFound('This model has no model card setup', { modelId, version })
-      }
-
-      modelCard = card
-    } else {
-      modelCard = await getModelCardRevision(req.user, modelId, version)
-    }
+    const modelCard = await getModelCardService(req.user, modelId, version)
 
     return res.json({ modelCard })
   },
